backend/requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
pandas>=2.2.2
numpy>=2.1.0
ollama==0.1.2
python-dotenv==1.0.0
pydantic==2.9.2
requests==2.31.0
nltk==3.8.1
beautifulsoup4==4.12.2
bcrypt==4.0.1
python-jose==3.3.0
passlib==1.7.4
python-multipart==0.0.6
PyYAML==6.0.2